refactor: move test into dedicated file
