+++ programs/dynamic-bonding-curve/src/state/fee.rs
@@ -143,106 +143,3 @@ impl FeeMode {
         })
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn test_fee_mode_output_token_base_to_quote() {
-        let fee_mode = FeeMode::get_fee_mode(
-            CollectFeeMode::OutputToken as u8,
-            TradeDirection::BaseToQuote,
-            false,
-        )
-        .unwrap();
-
-        assert_eq!(fee_mode.fees_on_input, false);
-        assert_eq!(fee_mode.fees_on_base_token, false);
-        assert_eq!(fee_mode.has_referral, false);
-    }
-
-    #[test]
-    fn test_fee_mode_output_token_quote_to_base() {
-        let fee_mode = FeeMode::get_fee_mode(
-            CollectFeeMode::OutputToken as u8,
-            TradeDirection::QuoteToBase,
-            true,
-        )
-        .unwrap();
-
-        assert_eq!(fee_mode.fees_on_input, false);
-        assert_eq!(fee_mode.fees_on_base_token, true);
-        assert_eq!(fee_mode.has_referral, true);
-    }
-
-    #[test]
-    fn test_fee_mode_quote_token_base_to_quote() {
-        let fee_mode = FeeMode::get_fee_mode(
-            CollectFeeMode::QuoteToken as u8,
-            TradeDirection::BaseToQuote,
-            false,
-        )
-        .unwrap();
-
-        assert_eq!(fee_mode.fees_on_input, false);
-        assert_eq!(fee_mode.fees_on_base_token, false);
-        assert_eq!(fee_mode.has_referral, false);
-    }
-
-    #[test]
-    fn test_fee_mode_quote_token_quote_to_base() {
-        let fee_mode = FeeMode::get_fee_mode(
-            CollectFeeMode::QuoteToken as u8,
-            TradeDirection::QuoteToBase,
-            true,
-        )
-        .unwrap();
-
-        assert_eq!(fee_mode.fees_on_input, true);
-        assert_eq!(fee_mode.fees_on_base_token, false);
-        assert_eq!(fee_mode.has_referral, true);
-    }
-
-    #[test]
-    fn test_invalid_collect_fee_mode() {
-        let result = FeeMode::get_fee_mode(
-            2, // Invalid mode
-            TradeDirection::QuoteToBase,
-            false,
-        );
-
-        assert!(result.is_err());
-    }
-
-    #[test]
-    fn test_fee_mode_default() {
-        let fee_mode = FeeMode::default();
-
-        assert_eq!(fee_mode.fees_on_input, false);
-        assert_eq!(fee_mode.fees_on_base_token, false);
-        assert_eq!(fee_mode.has_referral, false);
-    }
-
-    // Property-based test to ensure consistent behavior
-    #[test]
-    fn test_fee_mode_properties() {
-        // When trading BaseToQuote, fees should never be on input
-        let fee_mode = FeeMode::get_fee_mode(
-            CollectFeeMode::QuoteToken as u8,
-            TradeDirection::BaseToQuote,
-            true,
-        )
-        .unwrap();
-        assert_eq!(fee_mode.fees_on_input, false);
-
-        // When using QuoteToken mode, base_token should always be false
-        let fee_mode = FeeMode::get_fee_mode(
-            CollectFeeMode::QuoteToken as u8,
-            TradeDirection::QuoteToBase,
-            false,
-        )
-        .unwrap();
-        assert_eq!(fee_mode.fees_on_base_token, false);
-    }
-}
